move anari_library_debug headers to a structure that matches install dst
