+++ examples/example_device/ExampleDeviceDebugFactory.cpp
@@ -4,9 +4,7 @@
 // This file was generated by generate_debug_objects.py
 // Don't make changes to this directly
 
-#include "debug_device/DebugObject.h"
-namespace anari {
-namespace example_device {
+#include "anari/ext/debug/DebugObject.h"
 namespace anari {
 namespace debug_device {
 class ExampleDeviceDebugFactory : public ObjectFactory {
@@ -27,7 +25,7 @@ public:
    DebugObjectBase* new_group(DebugDevice *td, ANARIObject wh, ANARIObject h) override;
    DebugObjectBase* new_instance(DebugDevice *td, ANARIObject wh, ANARIObject h) override;
    DebugObjectBase* new_world(DebugDevice *td, ANARIObject wh, ANARIObject h) override;
-   DebugObjectBase* new_surface(DebugDevice *td, ANARIObject wh, ANARIObject h) override; 
+   DebugObjectBase* new_surface(DebugDevice *td, ANARIObject wh, ANARIObject h) override;
 };
 }
 }
@@ -279,7 +277,7 @@ class renderer_debug : public DebugObject<ANARI_RENDERER> {
       DebugObject::commit();
    }
 };
-class renderer_raycast : public DebugObject<ANARI_RENDERER> {
+class renderer_rayCast : public DebugObject<ANARI_RENDERER> {
    static int param_hash(const char *str) {
       static const uint32_t table[] = {0x6261000fu,0x0u,0x0u,0x0u,0x0u,0x0u,0x0u,0x0u,0x0u,0x0u,0x0u,0x0u,0x0u,0x0u,0x6a69001eu,0x64630010u,0x6c6b0011u,0x68670012u,0x73720013u,0x706f0014u,0x76750015u,0x6f6e0016u,0x65640017u,0x44430018u,0x706f0019u,0x6d6c001au,0x706f001bu,0x7372001cu,0x100001du,0x80000000u,0x7978001fu,0x66650020u,0x6d6c0021u,0x54530022u,0x62610023u,0x6e6d0024u,0x71700025u,0x6d6c0026u,0x66650027u,0x74730028u,0x1000029u,0x80000001u};
       uint32_t cur = 0x71620000u;
@@ -303,23 +301,23 @@ class renderer_raycast : public DebugObject<ANARI_RENDERER> {
       return -1;
    }
    public:
-   renderer_raycast(DebugDevice *td, ANARIObject wh, ANARIObject h): DebugObject(td, wh, h) { }
+   renderer_rayCast(DebugDevice *td, ANARIObject wh, ANARIObject h): DebugObject(td, wh, h) { }
    void setParameter(const char *paramname, ANARIDataType paramtype, const void *mem) {
       DebugObject::setParameter(paramname, paramtype, mem);
       int idx = param_hash(paramname);
       switch(idx) {
          case 0: { //backgroundColor
             ANARIDataType backgroundColor_types[] = {ANARI_FLOAT32_VEC4, ANARI_UNKNOWN};
-            check_type(ANARI_RENDERER, "raycast", paramname, paramtype, backgroundColor_types);
+            check_type(ANARI_RENDERER, "rayCast", paramname, paramtype, backgroundColor_types);
             return;
          }
          case 1: { //pixelSamples
             ANARIDataType pixelSamples_types[] = {ANARI_INT32, ANARI_UNKNOWN};
-            check_type(ANARI_RENDERER, "raycast", paramname, paramtype, pixelSamples_types);
+            check_type(ANARI_RENDERER, "rayCast", paramname, paramtype, pixelSamples_types);
             return;
          }
          default: // unknown param
-            unknown_parameter(ANARI_RENDERER, "raycast", paramname, paramtype);
+            unknown_parameter(ANARI_RENDERER, "rayCast", paramname, paramtype);
             return;
       }
    }
@@ -2780,7 +2778,7 @@ DebugObjectBase* ExampleDeviceDebugFactory::new_material(const char *name, Debug
    }
 }
 static int renderer_object_hash(const char *str) {
-   static const uint32_t table[] = {0x706f0013u,0x0u,0x0u,0x66650015u,0x0u,0x0u,0x0u,0x0u,0x0u,0x0u,0x0u,0x0u,0x0u,0x0u,0x0u,0x62610023u,0x0u,0x6261002du,0x64630056u,0x1000014u,0x80000000u,0x67620016u,0x7675001bu,0x0u,0x0u,0x0u,0x6261001eu,0x6867001cu,0x100001du,0x80000001u,0x7675001fu,0x6d6c0020u,0x75740021u,0x1000022u,0x80000002u,0x75740024u,0x69680025u,0x75740026u,0x73720027u,0x62610028u,0x64630029u,0x6665002au,0x7372002bu,0x100002cu,0x80000003u,0x7a79002eu,0x6444002fu,0x6a69004fu,0x0u,0x0u,0x0u,0x0u,0x0u,0x0u,0x0u,0x0u,0x0u,0x0u,0x0u,0x0u,0x0u,0x0u,0x0u,0x0u,0x0u,0x0u,0x0u,0x0u,0x0u,0x0u,0x0u,0x0u,0x0u,0x0u,0x0u,0x0u,0x0u,0x0u,0x62610052u,0x73720050u,0x1000051u,0x80000004u,0x74730053u,0x75740054u,0x1000055u,0x80000005u,0x6a690057u,0x77760058u,0x6a690059u,0x7473005au,0x100005bu,0x80000006u};
+   static const uint32_t table[] = {0x706f0013u,0x0u,0x0u,0x66650015u,0x0u,0x0u,0x0u,0x0u,0x0u,0x0u,0x0u,0x0u,0x0u,0x0u,0x0u,0x62610023u,0x0u,0x6261002du,0x64630038u,0x1000014u,0x80000000u,0x67620016u,0x7675001bu,0x0u,0x0u,0x0u,0x6261001eu,0x6867001cu,0x100001du,0x80000001u,0x7675001fu,0x6d6c0020u,0x75740021u,0x1000022u,0x80000002u,0x75740024u,0x69680025u,0x75740026u,0x73720027u,0x62610028u,0x64630029u,0x6665002au,0x7372002bu,0x100002cu,0x80000003u,0x7a79002eu,0x4543002fu,0x62610031u,0x6a690035u,0x74730032u,0x75740033u,0x1000034u,0x80000004u,0x73720036u,0x1000037u,0x80000005u,0x6a690039u,0x7776003au,0x6a69003bu,0x7473003cu,0x100003du,0x80000006u};
    uint32_t cur = 0x74610000u;
    for(int i = 0;cur!=0;++i) {
       uint32_t idx = cur&0xFFFFu;
@@ -2813,9 +2811,9 @@ DebugObjectBase* ExampleDeviceDebugFactory::new_renderer(const char *name, Debug
       case 3:
          return new renderer_pathtracer(td, wh, h);
       case 4:
-         return new renderer_rayDir(td, wh, h);
+         return new renderer_rayCast(td, wh, h);
       case 5:
-         return new renderer_raycast(td, wh, h);
+         return new renderer_rayDir(td, wh, h);
       case 6:
          return new renderer_scivis(td, wh, h);
       default:
@@ -2956,6 +2954,8 @@ DebugObjectBase* ExampleDeviceDebugFactory::new_world(DebugDevice *td, ANARIObje
 }
 }
 }
+namespace anari {
+namespace example_device {
 anari::debug_device::ObjectFactory* getDebugFactory() {
    static debug_device::ExampleDeviceDebugFactory f;
    return &f;
