+++ examples/generated_device_frontend/device/src/TreeDevice.cpp
@@ -9,7 +9,7 @@
 #include <cstdint>
 
 // debug interface
-#include "DebugObject.h"
+#include "anari/ext/debug/DebugObject.h"
 
 #include "RecursivePrint.h"
