+++ libs/debug_device/CMakeLists.txt
@@ -30,7 +30,8 @@ add_library(${PROJECT_NAME} SHARED
 
 target_include_directories(${PROJECT_NAME}
 PUBLIC
-  $<BUILD_INTERFACE:${CMAKE_CURRENT_LIST_DIR}>
+  $<BUILD_INTERFACE:${CMAKE_CURRENT_LIST_DIR}/include>
+  $<INSTALL_INTERFACE:${CMAKE_INSTALL_INCLUDEDIR}>
 )
 
 target_link_libraries(${PROJECT_NAME}
@@ -55,7 +56,6 @@ install(TARGETS ${PROJECT_NAME}
   ARCHIVE DESTINATION ${CMAKE_INSTALL_LIBDIR}
 )
 
-install(
-  FILES DebugObject.h debug_device_exports.h
-  DESTINATION ${CMAKE_INSTALL_INCLUDEDIR}/anari/ext/debug
+install(DIRECTORY ${CMAKE_CURRENT_LIST_DIR}/include/anari
+  DESTINATION ${CMAKE_INSTALL_INCLUDEDIR}
 )
