+++ libs/debug_device/GenericDeviceDebugFactory.cpp
@@ -4,9 +4,7 @@
 // This file was generated by generate_debug_objects.py
 // Don't make changes to this directly
 
-#include "debug_device/DebugObject.h"
-namespace anari {
-namespace debug_device {
+#include "anari/ext/debug/DebugObject.h"
 namespace anari {
 namespace debug_device {
 class GenericDeviceDebugFactory : public ObjectFactory {
@@ -27,7 +25,7 @@ public:
    DebugObjectBase* new_group(DebugDevice *td, ANARIObject wh, ANARIObject h) override;
    DebugObjectBase* new_instance(DebugDevice *td, ANARIObject wh, ANARIObject h) override;
    DebugObjectBase* new_world(DebugDevice *td, ANARIObject wh, ANARIObject h) override;
-   DebugObjectBase* new_surface(DebugDevice *td, ANARIObject wh, ANARIObject h) override; 
+   DebugObjectBase* new_surface(DebugDevice *td, ANARIObject wh, ANARIObject h) override;
 };
 }
 }
@@ -2646,6 +2644,8 @@ DebugObjectBase* GenericDeviceDebugFactory::new_world(DebugDevice *td, ANARIObje
 }
 }
 }
+namespace anari {
+namespace debug_device {
 anari::debug_device::ObjectFactory* getDebugFactory() {
    static debug_device::GenericDeviceDebugFactory f;
    return &f;
